+++ tests/Api/SetupSubscriptionTest.php
@@ -7,6 +7,7 @@ use Altapay\Response\SetupSubscriptionResponse as SetupSubscriptionResponse;
 use Altapay\Api\Subscription\SetupSubscription;
 use GuzzleHttp\Handler\MockHandler;
 use GuzzleHttp\Psr7\Response;
+use Altapay\Response\PaymentRequestResponse;
 
 class SetupSubscriptionTest extends AbstractApiTest
 {
@@ -15,7 +16,7 @@ class SetupSubscriptionTest extends AbstractApiTest
      */
     protected function getapi()
     {
-        $client = $this->getXmlClient(__DIR__ . '/Results/reservationoffixedamount.xml');
+        $client = $this->getXmlClient(__DIR__ . '/Results/reservation.xml');
 
         return (new SetupSubscription($this->getAuth()))
             ->setClient($client);
@@ -36,6 +37,13 @@ class SetupSubscriptionTest extends AbstractApiTest
         $api->setAmount(200.50);
         $api->setCurrency(957);
         $api->setShopOrderId('order id');
+        $api->setAgreement(
+            [
+                'id' => '232323232',
+                'agreement_type' => 'unscheduled',
+                'agreement[unscheduled]' => 'incremental'
+            ]
+        );
         $api->setSurcharge(155.23);
         $api->call();
     }
@@ -47,12 +55,19 @@ class SetupSubscriptionTest extends AbstractApiTest
         $api->setAmount(200.50);
         $api->setCurrency(957);
         $api->setShopOrderId('order id');
+        $api->setAgreement(
+            [
+                'id' => '232323232',
+                'agreement_type' => 'unscheduled',
+                'agreement[unscheduled]' => 'incremental'
+            ]
+        );
         $api->setSurcharge(155.23);
         $api->call();
         $request = $api->getRawRequest();
 
-        $this->assertSame($this->getExceptedUri('setupSubscription/'), $request->getUri()->getPath());
-        parse_str($request->getUri()->getQuery(), $parts);
+        $this->assertSame($this->getExceptedUri('setupSubscription'), $request->getUri()->getPath());
+        parse_str($request->getBody()->getContents(), $parts);
         $this->assertSame('my terminal', $parts['terminal']);
         $this->assertSame('order id', $parts['shop_orderid']);
         $this->assertSame('200.5', $parts['amount']);
@@ -67,11 +82,17 @@ class SetupSubscriptionTest extends AbstractApiTest
         $api->setAmount(200.50);
         $api->setCurrency(957);
         $api->setShopOrderId('order id');
-
+        $api->setAgreement(
+            [
+                'id' => '232323232',
+                'agreement_type' => 'unscheduled',
+                'agreement[unscheduled]' => 'incremental'
+            ]
+        );
         $response = $api->call();
 
         $this->assertInstanceOf(SetupSubscriptionResponse::class, $response);
         $this->assertSame('Success', $response->Result);
-        $this->assertCount(1, $response->Transactions);
+        $this->assertCount(1, (array)$response->Transactions);
     }
 }