+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.2.8] - 2022-11-01
+### Added
+- Supports API changes from 20221026
+- Add agreement setters for ChargeSubscription and ReserveSubscriptionCharge class
+- Add additional parameters for ReservationOfFixedAmount class
+- Add new property and setters for Terminal class
+- Add `subscription_model` as a new goods type for order line 
+### Fixes
+- Avoid throwing exception for missing setters or properties in response
+
 ## [3.2.7] - 2022-10-04
 ### Added
 - Add support for new 'Agreements Engine' parameters